Workaround an IllegalArgumentException error in world()
Something very strange happens with sf::st_make_valid() on
Binder Linux instance:

  LWGEOM_GEOS_nodeLines(): IllegalArgumentException: RobustDeterminant encountered non-finite numbers
  Error in CPL_make_valid(x) :
  IllegalArgumentException: RobustDeterminant encountered non-finite numbers

This does not happen on any of the testing Macs or any of the Linux
boxes I tested the function on. In fact, I had to add st_make_valid()
*because* I was getting errors with the "validity".

For now, I implemented a tryCatch solution. If an error happens,
st_make_valid() will not be applied.